SDA-2559: fix failing unit tests and refactor code
